Style pagination component for journeys
Add page-specific styling and ensure that links are targeted as
direct children to prevent inherited styles for the detail area.
The media query of 30rem is still mostly relevant, even if the
journey tends to break the detail into three lines, however these
viewports are in the tablet dead zone. Regardless, the content
renders fine, even if not exactly as desired.