Add styles for Replay steps component
This may be a temporary place for these to exist, as I'm not
sure if they should be bundled in each article's styles or not.
Regardless, a ton of styles are reused from the pagination
component, so it seemed easier to use this, instead.
However, aria and semantics may suffer due to this, so I'd
like to revisit this decision in the future.