Add partial matching to search
If there's an incompelte search term as the last term,
the search will attempt to use the token strategy with
a simulated IDF to partially match it to any document.
This allows for incomplete search terms, e.g. `javasc`
to be matched against any document proportionally
to the partial search term's length, providing a better
experience while typing a search term, instead of
only matching when the query is complete.